Apparently got the mod/rm /r and /[0-7] right for i386

+++ include/Asm/arch.h
@@ -155,7 +155,7 @@ typedef struct _ArchInstructionCall
 {
 	char const * name;
 	ArchOperand operands[3];
-	size_t operands_cnt;
+	uint32_t operands_cnt;
 } ArchInstructionCall;
 
 typedef struct _ArchRegister
